move types into seperate file, move actions into single file
