+++ src/actions/index.js
@@ -1,5 +1,6 @@
 import ACTIONS from './types';
 
+// WORKSPACE
 export const addWorkspace = workspace => {
     return {
         type: ACTIONS.ADD_WORKSPACE,
@@ -18,5 +19,26 @@ export const updateWorkspace = workspace => {
     return {
         type: ACTIONS.UPDATE_WORKSPACE,
         payload: workspace,
+    }
+}
+
+// EDIT WORKSPACE MODAL
+export function toggleWorkspaceModal() {
+    return {
+        type: ACTIONS.TOGGLE_WORKSPACE_MODAL,
+    }
+}
+
+// SELECTED WORKSPACE
+export const selectWorkspace = (space) => {
+    return {
+        type: ACTIONS.SELECT_WORKSPACE,
+        payload: space
+    }
+}
+
+export const unselectWorkspace = () => {
+    return {
+        type: ACTIONS.UNSELECT_WORKSPACE
     }
 }
