--- src/actions/toggleWorkspaceModal.js
@@ -1,7 +1,0 @@
-import ACTIONS from './types';
-
-export function toggleWorkspaceModal() {
-    return {
-        type: ACTIONS.TOGGLE_WORKSPACE_MODAL,
-    }
-}
