+++ src/component/Input/CreateWorkspace.jsx
@@ -2,8 +2,7 @@
 import React, { Component } from 'react';
 import styled from 'styled-components';
 import { connect } from 'react-redux';
-import { addWorkspace } from '../../actions/workspace';
-
+import { addWorkspace } from '../../actions';
 
 // Components
 import Input from './index';
